refactor: split jobs-worker.ts into per-job-type handler modules
Break the 1,154-line monolithic jobs-worker into focused modules:
- job-utils.ts: shared error classification, retry logic, embedding helpers, time windows
- job-handlers/embedding.ts: embed_segment, embed_item, embed_summary
- job-handlers/extraction.ts: extract_items, extract_entities
- job-handlers/conflict.ts: resolve_pending_conflicts_for_message, cleanup_resolved_conflicts
- job-handlers/cleanup.ts: check_contradictions, cleanup_stale_superseded_items
- job-handlers/summarization.ts: build_conversation_summary, refresh_weekly/monthly_summary
- job-handlers/backfill.ts: backfill, backfill_entity_relations
- job-handlers/index-maintenance.ts: rebuild_index, delete_qdrant_vectors

The main jobs-worker.ts retains the dispatch loop, worker lifecycle,
cleanup scheduling, and stale item sweep (~260 lines).
