+++ assistant/src/memory/job-handlers/index-maintenance.ts
@@ -1,0 +1,54 @@
+import { eq } from 'drizzle-orm';
+import { getLogger } from '../../util/logger.js';
+import { getDb } from '../db.js';
+import { enqueueMemoryJob, type MemoryJob } from '../jobs-store.js';
+import { asString, BackendUnavailableError } from '../job-utils.js';
+import { getQdrantClient } from '../qdrant-client.js';
+import { memoryEmbeddings, memoryItems, memorySegments, memorySummaries } from '../schema.js';
+
+const log = getLogger('memory-jobs-worker');
+
+export function rebuildIndexJob(): void {
+  const db = getDb();
+  db.run(/*sql*/ `DELETE FROM memory_segment_fts`);
+  db.run(/*sql*/ `
+    INSERT INTO memory_segment_fts(segment_id, text)
+    SELECT id, text FROM memory_segments
+  `);
+  db.delete(memoryEmbeddings).run();
+
+  const items = db
+    .select({ id: memoryItems.id })
+    .from(memoryItems)
+    .where(eq(memoryItems.status, 'active'))
+    .all();
+  for (const item of items) {
+    enqueueMemoryJob('embed_item', { itemId: item.id });
+  }
+
+  const summaries = db.select({ id: memorySummaries.id }).from(memorySummaries).all();
+  for (const summary of summaries) {
+    enqueueMemoryJob('embed_summary', { summaryId: summary.id });
+  }
+
+  const segments = db.select({ id: memorySegments.id }).from(memorySegments).all();
+  for (const segment of segments) {
+    enqueueMemoryJob('embed_segment', { segmentId: segment.id });
+  }
+}
+
+export async function deleteQdrantVectorsJob(job: MemoryJob): Promise<void> {
+  const targetType = asString(job.payload.targetType);
+  const targetId = asString(job.payload.targetId);
+  if (!targetType || !targetId) return;
+
+  let qdrant;
+  try {
+    qdrant = getQdrantClient();
+  } catch {
+    throw new BackendUnavailableError('Qdrant client not initialized');
+  }
+
+  await qdrant.deleteByTarget(targetType, targetId);
+  log.info({ targetType, targetId }, 'Retried Qdrant vector deletion succeeded');
+}
