+++ assistant/src/memory/jobs-worker.ts
@@ -1,157 +1,38 @@
-import Anthropic from '@anthropic-ai/sdk';
-import { and, asc, desc, eq, gt, gte, inArray, isNull, lt, ne, or } from 'drizzle-orm';
-import { v4 as uuid } from 'uuid';
 import type { AssistantConfig } from '../config/types.js';
 import { getConfig } from '../config/loader.js';
-import { estimateTextTokens } from '../context/token-estimator.js';
 import { getLogger } from '../util/logger.js';
-import {
-  readMessageCursorCheckpoint,
-  resetMessageCursorCheckpoint,
-  writeMessageCursorCheckpoint,
-} from './checkpoints.js';
-import { embedWithBackend, getMemoryBackendStatus } from './embedding-backend.js';
 import { getDb } from './db.js';
 import {
   claimMemoryJobs,
   completeMemoryJob,
   deferMemoryJob,
-  enqueueBackfillEntityRelationsJob,
   enqueueCleanupResolvedConflictsJob,
   enqueueCleanupStaleSupersededItemsJob,
-  enqueueMemoryJob,
   failMemoryJob,
   type MemoryJob,
   resetRunningJobsToPending,
 } from './jobs-store.js';
 import {
-  extractEntitiesWithLLM,
-  linkMemoryItemToEntity,
-  resolveEntityName,
-  upsertEntity,
-  upsertEntityRelation,
-} from './entity-extractor.js';
-import { indexMessageNow } from './indexer.js';
-import { checkContradictions } from './contradiction-checker.js';
-import { resolveConflictClarification } from './clarification-resolver.js';
-import { applyConflictResolution, listPendingConflictDetails } from './conflict-store.js';
-import { extractAndUpsertMemoryItemsForMessage } from './items-extractor.js';
-import { extractTextFromStoredMessageContent } from './message-content.js';
-import { getQdrantClient } from './qdrant-client.js';
-import {
-  memoryEmbeddings,
-  memoryItemConflicts,
-  memoryItemEntities,
-  memoryItems,
-  memoryItemSources,
-  memorySegments,
-  memorySummaries,
-  messages,
-} from './schema.js';
+  BackendUnavailableError,
+  classifyError,
+  retryDelayForAttempt,
+  RETRY_MAX_ATTEMPTS,
+} from './job-utils.js';
+
+// ── Per-job-type handlers ──────────────────────────────────────────
+
+import { embedSegmentJob, embedItemJob, embedSummaryJob } from './job-handlers/embedding.js';
+import { extractItemsJob, extractEntitiesJob } from './job-handlers/extraction.js';
+import { resolvePendingConflictsForMessageJob, cleanupResolvedConflictsJob } from './job-handlers/conflict.js';
+import { checkContradictionsJob, cleanupStaleSupersededItemsJob } from './job-handlers/cleanup.js';
+import { buildConversationSummaryJob, buildGlobalSummaryJob } from './job-handlers/summarization.js';
+import { backfillJob, backfillEntityRelationsJob } from './job-handlers/backfill.js';
+import { rebuildIndexJob, deleteQdrantVectorsJob } from './job-handlers/index-maintenance.js';
+
+// Re-export public utilities consumed by tests and other modules
+export { currentWeekWindow } from './job-utils.js';
 
 const log = getLogger('memory-jobs-worker');
-
-/** Sentinel error: the embedding backend is not configured yet. */
-class BackendUnavailableError extends Error {
-  constructor(reason: string) {
-    super(reason);
-    this.name = 'BackendUnavailableError';
-  }
-}
-
-// ── Error classification for LLM / API errors ─────────────────────
-
-type ErrorCategory = 'retryable' | 'fatal';
-
-const RETRY_BASE_DELAY_MS = 5_000;
-const RETRY_MAX_DELAY_MS = 5 * 60 * 1000;
-const RETRY_MAX_ATTEMPTS = 8;
-
-/**
- * Classify an error as retryable or fatal based on its HTTP status or type.
- *
- * Retryable: timeouts, 429 rate limits, 5xx server errors, connection errors.
- * Fatal: 400 bad request, 401 auth, 403 permission, other 4xx client errors.
- */
-function classifyError(err: unknown): ErrorCategory {
-  // Timeout errors from our own Promise.race wrappers
-  if (err instanceof Error && err.message.includes('timeout')) {
-    return 'retryable';
-  }
-
-  // SDK APIError subclasses (Anthropic and OpenAI share the same shape)
-  if (err != null && typeof err === 'object' && 'status' in err) {
-    const status = (err as { status?: number }).status;
-    if (typeof status === 'number') {
-      if (status === 429) return 'retryable';
-      if (status >= 500) return 'retryable';
-      // 400, 401, 403, 404, 409, 422, other 4xx → fatal
-      return 'fatal';
-    }
-    // No status (connection error) → retryable
-    return 'retryable';
-  }
-
-  // Parse HTTP status codes from error messages (e.g., "request failed (429): ...")
-  // Gemini and Ollama backends embed status codes in plain Error messages
-  if (err instanceof Error) {
-    const statusMatch = err.message.match(/\((\d{3})\)/);
-    if (statusMatch) {
-      const status = parseInt(statusMatch[1], 10);
-      if (status === 429) return 'retryable';
-      if (status >= 500) return 'retryable';
-      // 4xx client errors → fatal
-      if (status >= 400 && status < 500) return 'fatal';
-    }
-  }
-
-  // Connection/network errors without a status code
-  if (err instanceof Error && /ECONNREFUSED|ECONNRESET|ETIMEDOUT|ENETUNREACH|fetch failed/i.test(err.message)) {
-    return 'retryable';
-  }
-
-  // Unknown errors default to fatal to avoid infinite retry loops
-  return 'fatal';
-}
-
-/** Exponential backoff with full jitter: uniform random in [0, cap]. */
-function retryDelayForAttempt(attempts: number): number {
-  const cap = Math.min(RETRY_BASE_DELAY_MS * Math.pow(2, Math.max(0, attempts - 1)), RETRY_MAX_DELAY_MS);
-  return Math.random() * cap;
-}
-
-const BACKFILL_CHECKPOINT_KEY = 'memory:backfill:last_created_at';
-const BACKFILL_CHECKPOINT_ID_KEY = 'memory:backfill:last_message_id';
-const RELATION_BACKFILL_CHECKPOINT_KEY = 'memory:relation_backfill:last_created_at';
-const RELATION_BACKFILL_CHECKPOINT_ID_KEY = 'memory:relation_backfill:last_message_id';
-
-const SUMMARY_LLM_TIMEOUT_MS = 20_000;
-const SUMMARY_MAX_TOKENS = 800;
-const CLEANUP_BATCH_LIMIT = 250;
-
-const CONVERSATION_SUMMARY_SYSTEM_PROMPT = [
-  'You are a memory summarization system. Your job is to produce a compact, information-dense summary of a conversation.',
-  '',
-  'Guidelines:',
-  '- Focus on key facts, decisions, user preferences, and actionable information.',
-  '- Preserve concrete details: names, file paths, tool choices, technical decisions, constraints.',
-  '- Remove filler, pleasantries, and transient discussion that has no lasting value.',
-  '- Use concise bullet points grouped by topic.',
-  '- Target 400-600 tokens. Be dense but readable.',
-  '- If updating an existing summary with new data, merge new information and remove anything that was superseded.',
-].join('\n');
-
-const GLOBAL_SUMMARY_SYSTEM_PROMPT = [
-  'You are a memory summarization system. Your job is to synthesize a higher-level summary from multiple conversation summaries and memory items.',
-  '',
-  'Guidelines:',
-  '- Identify recurring themes, cross-cutting decisions, and persistent user preferences.',
-  '- Highlight the most important facts, active projects, and ongoing concerns.',
-  '- De-duplicate information that appears across multiple conversations.',
-  '- Use concise sections with bullet points.',
-  '- Target 400-600 tokens. Be dense but readable.',
-  '- If updating an existing summary with new data, merge new information and remove anything that was superseded.',
-].join('\n');
 
 export interface MemoryJobsWorker {
   runOnce(): Promise<number>;
@@ -251,6 +132,8 @@ export async function runMemoryJobsOnce(
   return processed;
 }
 
+// ── Job dispatch ───────────────────────────────────────────────────
+
 async function processJob(job: MemoryJob, config: AssistantConfig): Promise<void> {
   switch (job.type) {
     case 'embed_segment':
@@ -304,768 +187,6 @@ async function processJob(job: MemoryJob, config: AssistantConfig): Promise<void
     default:
       throw new Error(`Unknown memory job type: ${(job as { type: string }).type}`);
   }
-}
-
-async function embedSegmentJob(job: MemoryJob, config: AssistantConfig): Promise<void> {
-  const segmentId = asString(job.payload.segmentId);
-  if (!segmentId) return;
-  const db = getDb();
-  const segment = db
-    .select()
-    .from(memorySegments)
-    .where(eq(memorySegments.id, segmentId))
-    .get();
-  if (!segment) return;
-  await embedAndUpsert(config, 'segment', segment.id, segment.text, {
-    conversation_id: segment.conversationId,
-    message_id: segment.messageId,
-    created_at: segment.createdAt,
-  });
-}
-
-async function embedItemJob(job: MemoryJob, config: AssistantConfig): Promise<void> {
-  const itemId = asString(job.payload.itemId);
-  if (!itemId) return;
-  const db = getDb();
-  const item = db
-    .select()
-    .from(memoryItems)
-    .where(eq(memoryItems.id, itemId))
-    .get();
-  if (!item || item.status !== 'active') return;
-  const text = `<kind>${item.kind}</kind> ${item.subject}: ${item.statement}`;
-  await embedAndUpsert(config, 'item', item.id, text, {
-    kind: item.kind,
-    subject: item.subject,
-    status: item.status,
-    confidence: item.confidence,
-    created_at: item.firstSeenAt,
-    last_seen_at: item.lastSeenAt,
-  });
-}
-
-async function embedSummaryJob(job: MemoryJob, config: AssistantConfig): Promise<void> {
-  const summaryId = asString(job.payload.summaryId);
-  if (!summaryId) return;
-  const db = getDb();
-  const summary = db
-    .select()
-    .from(memorySummaries)
-    .where(eq(memorySummaries.id, summaryId))
-    .get();
-  if (!summary) return;
-  await embedAndUpsert(config, 'summary', summary.id, `[${summary.scope}] ${summary.summary}`, {
-    kind: summary.scope,
-    created_at: summary.startAt,
-    last_seen_at: summary.endAt,
-  });
-}
-
-async function extractItemsJob(job: MemoryJob): Promise<void> {
-  const messageId = asString(job.payload.messageId);
-  if (!messageId) return;
-  await extractAndUpsertMemoryItemsForMessage(messageId);
-  // Queue entity extraction for this message after items are extracted
-  const config = getConfig();
-  if (config.memory.entity.enabled) {
-    enqueueMemoryJob('extract_entities', { messageId });
-  }
-}
-
-async function extractEntitiesJob(job: MemoryJob, config: AssistantConfig): Promise<void> {
-  const messageId = asString(job.payload.messageId);
-  if (!messageId) return;
-
-  const db = getDb();
-  const message = db
-    .select({
-      id: messages.id,
-      content: messages.content,
-    })
-    .from(messages)
-    .where(eq(messages.id, messageId))
-    .get();
-  if (!message) return;
-
-  const text = extractTextFromStoredMessageContent(message.content);
-  if (text.trim().length < 15) return;
-
-  const extracted = await extractEntitiesWithLLM(text, config.memory.entity);
-  const entities = extracted.entities;
-  const relations = extracted.relations;
-  if (entities.length === 0 && relations.length === 0) return;
-
-  // Find all memory items linked to this message via memory_item_sources
-  const linkedItems = db
-    .select({ memoryItemId: memoryItemSources.memoryItemId })
-    .from(memoryItemSources)
-    .where(eq(memoryItemSources.messageId, messageId))
-    .all();
-  const itemIds = linkedItems.map((row) => row.memoryItemId);
-  const entityNameToId = new Map<string, string>();
-
-  for (const entity of entities) {
-    const entityId = upsertEntity(entity);
-    entityNameToId.set(entity.name.toLowerCase(), entityId);
-    for (const alias of entity.aliases) {
-      entityNameToId.set(alias.toLowerCase(), entityId);
-    }
-    // Link all memory items from this message to the entity
-    for (const itemId of itemIds) {
-      linkMemoryItemToEntity(itemId, entityId);
-    }
-  }
-
-  const relationTelemetry = {
-    attempted: 0,
-    parsed: relations.length,
-    persisted: 0,
-    dropped: 0,
-  };
-
-  if (config.memory.entity.extractRelations.enabled && relations.length > 0) {
-    const seenRelationKeys = new Set<string>();
-    for (const relation of relations) {
-      relationTelemetry.attempted += 1;
-      const sourceLookup = relation.sourceEntityName.toLowerCase();
-      const targetLookup = relation.targetEntityName.toLowerCase();
-      const sourceEntityId = entityNameToId.get(sourceLookup) ?? resolveEntityName(relation.sourceEntityName);
-      const targetEntityId = entityNameToId.get(targetLookup) ?? resolveEntityName(relation.targetEntityName);
-      if (!sourceEntityId || !targetEntityId || sourceEntityId === targetEntityId) {
-        relationTelemetry.dropped += 1;
-        continue;
-      }
-
-      const dedupeKey = `${sourceEntityId}|${targetEntityId}|${relation.relation}`;
-      if (seenRelationKeys.has(dedupeKey)) continue;
-      seenRelationKeys.add(dedupeKey);
-
-      upsertEntityRelation({
-        sourceEntityId,
-        targetEntityId,
-        relation: relation.relation,
-        evidence: relation.evidence,
-      });
-      relationTelemetry.persisted += 1;
-    }
-  }
-
-  log.debug({
-    messageId,
-    entityCount: entities.length,
-    linkedItems: itemIds.length,
-    relationAttempts: relationTelemetry.attempted,
-    relationParsed: relationTelemetry.parsed,
-    relationPersisted: relationTelemetry.persisted,
-    relationDropped: relationTelemetry.dropped,
-  }, 'Extracted entity graph from message');
-}
-
-async function checkContradictionsJob(job: MemoryJob): Promise<void> {
-  const itemId = asString(job.payload.itemId);
-  if (!itemId) return;
-  await checkContradictions(itemId);
-}
-
-async function resolvePendingConflictsForMessageJob(job: MemoryJob, config: AssistantConfig): Promise<void> {
-  if (!config.memory.conflicts.enabled) return;
-  const messageId = asString(job.payload.messageId);
-  if (!messageId) return;
-  const scopeId = asString(job.payload.scopeId) ?? 'default';
-  const db = getDb();
-  const message = db
-    .select({
-      id: messages.id,
-      role: messages.role,
-      content: messages.content,
-      createdAt: messages.createdAt,
-    })
-    .from(messages)
-    .where(eq(messages.id, messageId))
-    .get();
-  if (!message || message.role !== 'user') return;
-
-  const userMessage = extractTextFromStoredMessageContent(message.content).trim();
-  if (userMessage.length === 0) return;
-
-  const pending = listPendingConflictDetails(scopeId, 25);
-  const eligible = pending.filter((conflict) => conflict.createdAt <= message.createdAt);
-  if (eligible.length === 0) return;
-
-  let resolvedCount = 0;
-  for (const conflict of eligible) {
-    const resolution = await resolveConflictClarification(
-      {
-        existingStatement: conflict.existingStatement,
-        candidateStatement: conflict.candidateStatement,
-        userMessage,
-      },
-      { timeoutMs: config.memory.conflicts.resolverLlmTimeoutMs },
-    );
-    if (resolution.resolution === 'still_unclear') continue;
-    const resolved = applyConflictResolution({
-      conflictId: conflict.id,
-      resolution: resolution.resolution,
-      mergedStatement: resolution.resolution === 'merge' ? resolution.resolvedStatement : null,
-      resolutionNote: `Background message resolver (${resolution.strategy}): ${resolution.explanation}`,
-    });
-    if (resolved) resolvedCount += 1;
-  }
-
-  log.debug({
-    messageId,
-    scopeId,
-    pendingConflicts: pending.length,
-    eligibleConflicts: eligible.length,
-    resolvedConflicts: resolvedCount,
-  }, 'Processed pending conflict resolution job');
-}
-
-function cleanupResolvedConflictsJob(job: MemoryJob, config: AssistantConfig): void {
-  const db = getDb();
-  const retentionMs = asPositiveMs(job.payload.retentionMs) ?? config.memory.cleanup.resolvedConflictRetentionMs;
-  const cutoff = Date.now() - retentionMs;
-  const stale = db
-    .select({ id: memoryItemConflicts.id })
-    .from(memoryItemConflicts)
-    .where(and(
-      ne(memoryItemConflicts.status, 'pending_clarification'),
-      lt(memoryItemConflicts.resolvedAt, cutoff),
-    ))
-    .orderBy(asc(memoryItemConflicts.resolvedAt), asc(memoryItemConflicts.id))
-    .limit(CLEANUP_BATCH_LIMIT)
-    .all();
-  if (stale.length === 0) return;
-
-  const ids = stale.map((row) => row.id);
-  db.delete(memoryItemConflicts)
-    .where(inArray(memoryItemConflicts.id, ids))
-    .run();
-  if (stale.length === CLEANUP_BATCH_LIMIT) {
-    enqueueMemoryJob('cleanup_resolved_conflicts', { retentionMs });
-  }
-
-  log.debug({
-    removedConflicts: stale.length,
-    retentionMs,
-    cutoff,
-  }, 'Cleaned up resolved memory conflicts');
-}
-
-function cleanupStaleSupersededItemsJob(job: MemoryJob, config: AssistantConfig): void {
-  const db = getDb();
-  const retentionMs = asPositiveMs(job.payload.retentionMs) ?? config.memory.cleanup.supersededItemRetentionMs;
-  const cutoff = Date.now() - retentionMs;
-  const stale = db
-    .select({ id: memoryItems.id })
-    .from(memoryItems)
-    .where(and(
-      eq(memoryItems.status, 'superseded'),
-      lt(memoryItems.invalidAt, cutoff),
-    ))
-    .orderBy(asc(memoryItems.invalidAt), asc(memoryItems.id))
-    .limit(CLEANUP_BATCH_LIMIT)
-    .all();
-  if (stale.length === 0) return;
-
-  const ids = stale.map((row) => row.id);
-  db.delete(memoryItemEntities)
-    .where(inArray(memoryItemEntities.memoryItemId, ids))
-    .run();
-  db.delete(memoryEmbeddings)
-    .where(and(
-      eq(memoryEmbeddings.targetType, 'item'),
-      inArray(memoryEmbeddings.targetId, ids),
-    ))
-    .run();
-  db.delete(memoryItems)
-    .where(inArray(memoryItems.id, ids))
-    .run();
-  if (stale.length === CLEANUP_BATCH_LIMIT) {
-    enqueueMemoryJob('cleanup_stale_superseded_items', { retentionMs });
-  }
-
-  log.debug({
-    removedItems: stale.length,
-    retentionMs,
-    cutoff,
-  }, 'Cleaned up stale superseded memory items');
-}
-
-async function buildConversationSummaryJob(job: MemoryJob, config: AssistantConfig): Promise<void> {
-  const conversationId = asString(job.payload.conversationId);
-  if (!conversationId) return;
-  const db = getDb();
-  const rows = db
-    .select()
-    .from(memorySegments)
-    .where(eq(memorySegments.conversationId, conversationId))
-    .orderBy(desc(memorySegments.createdAt))
-    .limit(40)
-    .all();
-  if (rows.length === 0) return;
-
-  const existing = db
-    .select()
-    .from(memorySummaries)
-    .where(and(
-      eq(memorySummaries.scope, 'conversation'),
-      eq(memorySummaries.scopeKey, conversationId),
-    ))
-    .get();
-
-  // Build segment text for LLM input (chronological order)
-  const segmentTexts = rows
-    .slice(0, 30)
-    .reverse()
-    .map((row) => `[${row.role}] ${truncate(row.text, 400)}`)
-    .join('\n\n');
-
-  const summaryText = await summarizeWithLLM(
-    config,
-    CONVERSATION_SUMMARY_SYSTEM_PROMPT,
-    existing?.summary ?? null,
-    segmentTexts,
-    'conversation',
-  );
-
-  const now = Date.now();
-  const summaryId = existing?.id ?? uuid();
-  const nextVersion = (existing?.version ?? 0) + 1;
-  if (existing) {
-    db.update(memorySummaries)
-      .set({
-        summary: summaryText,
-        tokenEstimate: estimateTextTokens(summaryText),
-        version: nextVersion,
-        startAt: rows[rows.length - 1].createdAt,
-        endAt: rows[0].createdAt,
-        updatedAt: now,
-      })
-      .where(eq(memorySummaries.id, existing.id))
-      .run();
-  } else {
-    db.insert(memorySummaries).values({
-      id: summaryId,
-      scope: 'conversation',
-      scopeKey: conversationId,
-      summary: summaryText,
-      tokenEstimate: estimateTextTokens(summaryText),
-      version: nextVersion,
-      startAt: rows[rows.length - 1].createdAt,
-      endAt: rows[0].createdAt,
-      createdAt: now,
-      updatedAt: now,
-    }).run();
-  }
-  enqueueMemoryJob('embed_summary', { summaryId });
-}
-
-async function buildGlobalSummaryJob(scope: 'weekly_global' | 'monthly_global', config: AssistantConfig): Promise<void> {
-  const db = getDb();
-  const now = new Date();
-  const { startMs, endMs, scopeKey } = scope === 'weekly_global'
-    ? currentWeekWindow(now)
-    : currentMonthWindow(now);
-
-  const items = db
-    .select()
-    .from(memoryItems)
-    .where(and(
-      eq(memoryItems.status, 'active'),
-      isNull(memoryItems.invalidAt),
-      gte(memoryItems.lastSeenAt, startMs),
-      lt(memoryItems.lastSeenAt, endMs),
-    ))
-    .orderBy(desc(memoryItems.lastSeenAt))
-    .limit(80)
-    .all();
-
-  // Gather conversation summaries from this period for higher-level synthesis
-  const convSummaries = db
-    .select()
-    .from(memorySummaries)
-    .where(and(
-      eq(memorySummaries.scope, 'conversation'),
-      gte(memorySummaries.endAt, startMs),
-      lt(memorySummaries.startAt, endMs),
-    ))
-    .orderBy(desc(memorySummaries.endAt))
-    .limit(20)
-    .all();
-
-  if (items.length === 0 && convSummaries.length === 0) return;
-
-  // Build input for LLM: conversation summaries + active items
-  const parts: string[] = [];
-  if (convSummaries.length > 0) {
-    parts.push('## Conversation Summaries');
-    for (const cs of convSummaries) {
-      parts.push(`### ${cs.scopeKey}\n${truncate(cs.summary, 600)}`);
-    }
-  }
-  if (items.length > 0) {
-    parts.push('## Active Memory Items');
-    for (const item of items.slice(0, 40)) {
-      parts.push(`- [${item.kind}] ${item.subject}: ${truncate(item.statement, 180)}`);
-    }
-  }
-  const inputText = parts.join('\n\n');
-
-  const existing = db
-    .select()
-    .from(memorySummaries)
-    .where(and(
-      eq(memorySummaries.scope, scope),
-      eq(memorySummaries.scopeKey, scopeKey),
-    ))
-    .get();
-
-  const label = scope === 'weekly_global' ? 'weekly' : 'monthly';
-  const summaryText = await summarizeWithLLM(
-    config,
-    GLOBAL_SUMMARY_SYSTEM_PROMPT,
-    existing?.summary ?? null,
-    inputText,
-    label,
-  );
-
-  const ts = Date.now();
-  const summaryId = existing?.id ?? uuid();
-  const nextVersion = (existing?.version ?? 0) + 1;
-  if (existing) {
-    db.update(memorySummaries)
-      .set({
-        summary: summaryText,
-        tokenEstimate: estimateTextTokens(summaryText),
-        version: nextVersion,
-        startAt: startMs,
-        endAt: endMs,
-        updatedAt: ts,
-      })
-      .where(eq(memorySummaries.id, existing.id))
-      .run();
-  } else {
-    db.insert(memorySummaries).values({
-      id: summaryId,
-      scope,
-      scopeKey,
-      summary: summaryText,
-      tokenEstimate: estimateTextTokens(summaryText),
-      version: nextVersion,
-      startAt: startMs,
-      endAt: endMs,
-      createdAt: ts,
-      updatedAt: ts,
-    }).run();
-  }
-  enqueueMemoryJob('embed_summary', { summaryId });
-}
-
-async function summarizeWithLLM(
-  config: AssistantConfig,
-  systemPrompt: string,
-  existingSummary: string | null,
-  newContent: string,
-  label: string,
-): Promise<string> {
-  const summarizationConfig = config.memory.summarization;
-  if (!summarizationConfig.useLLM) {
-    log.debug({ label }, 'LLM summarization disabled, using fallback');
-    return buildFallbackSummary(existingSummary, newContent, label);
-  }
-
-  const apiKey = config.apiKeys.anthropic ?? process.env.ANTHROPIC_API_KEY;
-  if (!apiKey) {
-    log.debug({ label }, 'No Anthropic API key available for summarization, using fallback');
-    return buildFallbackSummary(existingSummary, newContent, label);
-  }
-
-  const userParts: string[] = [];
-  if (existingSummary) {
-    userParts.push(
-      '### Existing Summary (update with new data, keep what is still relevant, remove superseded info)',
-      existingSummary,
-      '',
-    );
-  }
-  userParts.push('### New Data', newContent);
-
-  try {
-    const client = new Anthropic({ apiKey });
-    const response = await Promise.race([
-      client.messages.create({
-        model: summarizationConfig.model,
-        max_tokens: SUMMARY_MAX_TOKENS,
-        system: systemPrompt,
-        messages: [{ role: 'user' as const, content: userParts.join('\n') }],
-      }),
-      new Promise<never>((_, reject) =>
-        setTimeout(() => reject(new Error('Summarization LLM timeout')), SUMMARY_LLM_TIMEOUT_MS),
-      ),
-    ]);
-
-    const textBlock = response.content.find((b) => b.type === 'text');
-    if (textBlock && textBlock.type === 'text' && textBlock.text.trim().length > 0) {
-      log.debug(
-        { label, inputTokens: response.usage.input_tokens, outputTokens: response.usage.output_tokens },
-        'LLM summarization completed',
-      );
-      return textBlock.text.trim();
-    }
-
-    log.warn({ label }, 'LLM summarization returned empty text, using fallback');
-    return buildFallbackSummary(existingSummary, newContent, label);
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.warn({ err: message, label }, 'LLM summarization failed, using fallback');
-    return buildFallbackSummary(existingSummary, newContent, label);
-  }
-}
-
-function buildFallbackSummary(_existingSummary: string | null, newContent: string, label: string): string {
-  const lines = newContent.split('\n').filter((l) => l.trim().length > 0);
-  const snippets = lines.slice(0, 20).map((l) => `- ${truncate(l.trim(), 180)}`);
-  const parts: string[] = [`${label} summary`, '', ...snippets];
-  return parts.join('\n');
-}
-
-function backfillJob(job: MemoryJob, config: AssistantConfig): void {
-  const db = getDb();
-  const force = job.payload.force === true;
-  if (force) {
-    resetMessageCursorCheckpoint(BACKFILL_CHECKPOINT_KEY, BACKFILL_CHECKPOINT_ID_KEY);
-  }
-
-  const cursor = readMessageCursorCheckpoint(BACKFILL_CHECKPOINT_KEY, BACKFILL_CHECKPOINT_ID_KEY);
-  const batch = db
-    .select()
-    .from(messages)
-    .where(or(
-      gt(messages.createdAt, cursor.createdAt),
-      and(eq(messages.createdAt, cursor.createdAt), gt(messages.id, cursor.messageId)),
-    ))
-    .orderBy(asc(messages.createdAt), asc(messages.id))
-    .limit(200)
-    .all();
-
-  if (batch.length > 0) {
-    for (const message of batch) {
-      indexMessageNow({
-        messageId: message.id,
-        conversationId: message.conversationId,
-        role: message.role,
-        content: message.content,
-        createdAt: message.createdAt,
-      }, config.memory);
-    }
-    const lastMessage = batch[batch.length - 1];
-    writeMessageCursorCheckpoint(BACKFILL_CHECKPOINT_KEY, BACKFILL_CHECKPOINT_ID_KEY, {
-      createdAt: lastMessage.createdAt,
-      messageId: lastMessage.id,
-    });
-  }
-
-  if (batch.length === 200) {
-    enqueueMemoryJob('backfill', {});
-  } else if (config.memory.entity.enabled && config.memory.entity.extractRelations.enabled) {
-    // Enqueue after the terminal batch (including an empty batch when total
-    // messages are an exact multiple of 200) so the relation backfill does not
-    // overlap with messages the normal backfill already covered via
-    // indexMessageNow → extract_items → extract_entities.
-    enqueueBackfillEntityRelationsJob();
-  }
-}
-
-function backfillEntityRelationsJob(job: MemoryJob, config: AssistantConfig): void {
-  if (!config.memory.entity.enabled) return;
-  if (!config.memory.entity.extractRelations.enabled) return;
-
-  const force = job.payload.force === true;
-  if (force) {
-    resetMessageCursorCheckpoint(RELATION_BACKFILL_CHECKPOINT_KEY, RELATION_BACKFILL_CHECKPOINT_ID_KEY);
-  }
-
-  const db = getDb();
-  const cursor = readMessageCursorCheckpoint(
-    RELATION_BACKFILL_CHECKPOINT_KEY,
-    RELATION_BACKFILL_CHECKPOINT_ID_KEY,
-  );
-  const batchSize = Math.max(1, config.memory.entity.extractRelations.backfillBatchSize);
-
-  const afterCursor = or(
-    gt(messages.createdAt, cursor.createdAt),
-    and(eq(messages.createdAt, cursor.createdAt), gt(messages.id, cursor.messageId)),
-  );
-
-  // Honor extractFromAssistant config — same role filter as indexMessageNow
-  const roleFilter = config.memory.extraction.extractFromAssistant
-    ? undefined
-    : ne(messages.role, 'assistant');
-
-  const conditions = roleFilter
-    ? and(afterCursor, roleFilter)
-    : afterCursor;
-
-  const batch = db
-    .select({ id: messages.id, role: messages.role, createdAt: messages.createdAt })
-    .from(messages)
-    .where(conditions)
-    .orderBy(asc(messages.createdAt), asc(messages.id))
-    .limit(batchSize)
-    .all();
-  if (batch.length === 0) return;
-
-  for (const message of batch) {
-    enqueueMemoryJob('extract_entities', { messageId: message.id });
-  }
-
-  const lastMessage = batch[batch.length - 1];
-  writeMessageCursorCheckpoint(RELATION_BACKFILL_CHECKPOINT_KEY, RELATION_BACKFILL_CHECKPOINT_ID_KEY, {
-    createdAt: lastMessage.createdAt,
-    messageId: lastMessage.id,
-  });
-
-  if (batch.length === batchSize) {
-    enqueueBackfillEntityRelationsJob();
-  }
-
-  log.debug({
-    queuedExtractEntityJobs: batch.length,
-    batchSize,
-    lastCreatedAt: lastMessage.createdAt,
-    lastMessageId: lastMessage.id,
-  }, 'Queued relation backfill batch');
-}
-
-function rebuildIndexJob(): void {
-  const db = getDb();
-  db.run(/*sql*/ `DELETE FROM memory_segment_fts`);
-  db.run(/*sql*/ `
-    INSERT INTO memory_segment_fts(segment_id, text)
-    SELECT id, text FROM memory_segments
-  `);
-  db.delete(memoryEmbeddings).run();
-
-  const items = db
-    .select({ id: memoryItems.id })
-    .from(memoryItems)
-    .where(eq(memoryItems.status, 'active'))
-    .all();
-  for (const item of items) {
-    enqueueMemoryJob('embed_item', { itemId: item.id });
-  }
-
-  const summaries = db.select({ id: memorySummaries.id }).from(memorySummaries).all();
-  for (const summary of summaries) {
-    enqueueMemoryJob('embed_summary', { summaryId: summary.id });
-  }
-
-  const segments = db.select({ id: memorySegments.id }).from(memorySegments).all();
-  for (const segment of segments) {
-    enqueueMemoryJob('embed_segment', { segmentId: segment.id });
-  }
-}
-
-async function deleteQdrantVectorsJob(job: MemoryJob): Promise<void> {
-  const targetType = asString(job.payload.targetType);
-  const targetId = asString(job.payload.targetId);
-  if (!targetType || !targetId) return;
-
-  let qdrant;
-  try {
-    qdrant = getQdrantClient();
-  } catch {
-    throw new BackendUnavailableError('Qdrant client not initialized');
-  }
-
-  await qdrant.deleteByTarget(targetType, targetId);
-  log.info({ targetType, targetId }, 'Retried Qdrant vector deletion succeeded');
-}
-
-async function embedAndUpsert(
-  config: AssistantConfig,
-  targetType: 'segment' | 'item' | 'summary',
-  targetId: string,
-  text: string,
-  extraPayload?: Record<string, unknown>,
-): Promise<void> {
-  const status = getMemoryBackendStatus(config);
-  if (!status.provider) {
-    throw new BackendUnavailableError(
-      `Embedding backend unavailable (${status.reason ?? 'no provider'})`,
-    );
-  }
-
-  const embedded = await embedWithBackend(config, [text]);
-  const vector = embedded.vectors[0];
-  if (!vector) return;
-
-  let qdrant;
-  try {
-    qdrant = getQdrantClient();
-  } catch {
-    throw new BackendUnavailableError('Qdrant client not initialized');
-  }
-
-  try {
-    const now = Date.now();
-    await qdrant.upsert(targetType, targetId, vector, {
-      text,
-      created_at: (extraPayload?.created_at as number) ?? now,
-      ...(extraPayload as Record<string, unknown> | undefined),
-    });
-  } catch (err) {
-    log.warn({ err, targetType, targetId }, 'Failed to upsert embedding to Qdrant');
-    throw err;
-  }
-}
-
-function asString(value: unknown): string | null {
-  return typeof value === 'string' && value.length > 0 ? value : null;
-}
-
-function asPositiveMs(value: unknown): number | null {
-  if (typeof value !== 'number') return null;
-  if (!Number.isFinite(value) || value <= 0) return null;
-  return Math.floor(value);
-}
-
-function truncate(text: string, max: number): string {
-  if (text.length <= max) return text;
-  return `${text.slice(0, max - 3)}...`;
-}
-
-export function currentWeekWindow(now: Date): { scopeKey: string; startMs: number; endMs: number } {
-  const day = (now.getUTCDay() + 6) % 7;
-  const start = new Date(Date.UTC(
-    now.getUTCFullYear(),
-    now.getUTCMonth(),
-    now.getUTCDate() - day,
-    0,
-    0,
-    0,
-    0,
-  ));
-  const end = new Date(start);
-  end.setUTCDate(end.getUTCDate() + 7);
-  const scopeKey = `${start.getUTCFullYear()}-W${weekNumber(start).toString().padStart(2, '0')}`;
-  return { scopeKey, startMs: start.getTime(), endMs: end.getTime() };
-}
-
-function currentMonthWindow(now: Date): { scopeKey: string; startMs: number; endMs: number } {
-  const start = new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth(), 1, 0, 0, 0, 0));
-  const end = new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth() + 1, 1, 0, 0, 0, 0));
-  const scopeKey = `${start.getUTCFullYear()}-${String(start.getUTCMonth() + 1).padStart(2, '0')}`;
-  return { scopeKey, startMs: start.getTime(), endMs: end.getTime() };
-}
-
-function weekNumber(date: Date): number {
-  const tmp = new Date(Date.UTC(date.getUTCFullYear(), date.getUTCMonth(), date.getUTCDate()));
-  const dayNum = tmp.getUTCDay() || 7;
-  tmp.setUTCDate(tmp.getUTCDate() + 4 - dayNum);
-  const yearStart = new Date(Date.UTC(tmp.getUTCFullYear(), 0, 1));
-  return Math.ceil((((tmp.getTime() - yearStart.getTime()) / 86400000) + 1) / 7);
 }
 
 // ── Cleanup scheduling ─────────────────────────────────────────────
